+++ backend/src/services/api.rs
@@ -4,10 +4,6 @@ use super::{auth::HttpClient, prelude::*};
 // think of better method of returning info
 // another thread that constantly updates info in period and cache the results?
 
-// TODO: For methods, that requires information from all nodes (/disks/count, /nodes/rps, etc.),
-// think of better method of returning info
-// another thread that constantly updates info in period and cache the results?
-
 /// Returns count of Physical Disks per status
 #[cfg_attr(all(feature = "swagger", debug_assertions),
     utoipa::path(
@@ -400,6 +396,146 @@ pub async fn get_node_info(
     Ok(Json(node))
 }
 
+async fn process_replica(
+    client: &HttpBobClient,
+    replica: dto::Replica,
+    disks: &HashMap<DiskName, IsActive>,
+    nodes: &HashMap<&NodeName, &dto::Node>,
+) -> Replica {
+    let mut status = ReplicaStatus::Good;
+    if let Some(disk_state) = disks.get(&replica.disk) {
+        if !disk_state {
+            status = ReplicaStatus::Offline {
+                problems: vec![ReplicaProblem::DiskUnavailable],
+            };
+        }
+    } else {
+        status = ReplicaStatus::Offline {
+            problems: vec![ReplicaProblem::DiskUnavailable],
+        };
+    }
+
+    if let Some(node) = nodes.get(&replica.node) {
+        if !is_node_online(client, node).await {
+            status = match status {
+                ReplicaStatus::Good => ReplicaStatus::Offline {
+                    problems: vec![ReplicaProblem::DiskUnavailable],
+                },
+                ReplicaStatus::Offline { mut problems } => {
+                    problems.push(ReplicaProblem::NodeUnavailable);
+                    ReplicaStatus::Offline { problems }
+                }
+            }
+        }
+    } else {
+        status = match status {
+            ReplicaStatus::Good => ReplicaStatus::Offline {
+                problems: vec![ReplicaProblem::DiskUnavailable],
+            },
+            ReplicaStatus::Offline { mut problems } => {
+                problems.push(ReplicaProblem::NodeUnavailable);
+                ReplicaStatus::Offline { problems }
+            }
+        }
+    }
+
+    Replica {
+        node: replica.node,
+        disk: replica.disk,
+        path: replica.path,
+        status,
+    }
+}
+
+async fn is_node_online(client: &HttpBobClient, node: &dto::Node) -> bool {
+    (client.probe_secondary(&node.name).await).map_or(false, |code| code == StatusCode::OK)
+}
+
+fn proccess_disks(
+    disks: &[dto::DiskState],
+    space: &dto::SpaceInfo,
+    metrics: &dto::MetricsSnapshotModel,
+) -> Vec<Disk> {
+    let mut res_disks = vec![];
+    let mut visited_disks = HashSet::new();
+    for disk in disks {
+        if !visited_disks.insert(disk.name.clone()) {
+            tracing::warn!(
+                "disk {} with path {} duplicated, skipping...",
+                disk.name,
+                disk.path
+            );
+            continue;
+        }
+        res_disks.push(Disk {
+            name: disk.name.clone(),
+            path: disk.path.clone(),
+            status: DiskStatus::from_space_info(space, &disk.name),
+            total_space: space.total_disk_space_bytes,
+            used_space: space
+                .occupied_disk_space_by_disk
+                .get(&disk.name.clone())
+                .copied()
+                .unwrap_or_default(),
+            iops: metrics
+                .metrics
+                .get(&format!("hardware.disks.{:?}_iops", disk.name))
+                .cloned()
+                .unwrap_or_default()
+                .value,
+        });
+    }
+
+    res_disks
+}
+
+// Bad function, 6 args :(
+async fn process_vdisks_for_node(
+    client: HttpBobClient,
+    virt_disks: &[dto::VDisk],
+    node_name: NodeName,
+    all_disks: &HashMap<DiskName, IsActive>,
+    nodes: &HashMap<&NodeName, &dto::Node>,
+    partitions_count_on_vdisk: &HashMap<i32, usize>,
+) -> Vec<VDisk> {
+    let mut res_replicas = vec![];
+    let mut res_vdisks = vec![];
+    for (vdisk, replicas) in virt_disks
+        .iter()
+        .filter_map(|vdisk| vdisk.replicas.as_ref().map(|repl| (vdisk, repl)))
+        .filter(|(_, replicas)| replicas.iter().any(|replica| replica.node == node_name))
+    {
+        for replica in replicas {
+            res_replicas.push((
+                vdisk.id,
+                process_replica(&client, replica.clone(), all_disks, nodes).await,
+            ));
+        }
+        res_vdisks.push(VDisk {
+            id: vdisk.id as u64,
+            status: if res_replicas
+                .iter()
+                .any(|(_, replica)| matches!(replica.status, ReplicaStatus::Offline { .. }))
+            {
+                VDiskStatus::Bad
+            } else {
+                VDiskStatus::Good
+            },
+            partition_count: partitions_count_on_vdisk
+                .get(&vdisk.id)
+                .copied()
+                .unwrap_or_default() as u64,
+            replicas: res_replicas
+                .iter()
+                .filter(|(id, _)| id == &vdisk.id)
+                .map(|(_, replica)| replica.clone())
+                .collect(),
+        });
+    }
+
+    res_vdisks
+}
+
 /// Get Raw Metrics from Node
 ///
 /// # Errors
@@ -461,6 +597,137 @@ pub async fn raw_configuration_by_node(
         )
         .await?,
     ))
+}
+
+/// Get Detailed Information on Node
+///
+/// # Errors
+///
+/// This function will return an error if the server was unable to get node'a client
+/// or one of the requests to get information from the node fails
+#[cfg_attr(feature = "swagger", utoipa::path(
+        get,
+        context_path = ApiV1::to_path(),
+        path = "/nodes/{node_name}/detailed",
+        params (
+            ("id", description = "Node's ID")
+        ),
+        responses(
+            (status = 200, body = DetailedNode, content_type = "application/json", description = "Detailed Node information"),
+
+            (status = 401, description = "Unauthorized"),
+            (status = 404, description = "Node Not Found")
+        ),
+        security(("api_key" = []))
+    ))]
+pub async fn get_detailed_node_info(
+    Extension(client): Extension<HttpBobClient>,
+    Path(node_name): Path<NodeName>,
+) -> AxumResult<Json<DetailedNode>> {
+    let mut all_disks: FuturesUnordered<_> = client
+        .cluster()
+        .map(move |node| {
+            let handle = node.clone();
+            tokio::spawn(async move { handle.get_disks().await })
+        })
+        .collect();
+
+    let node_client = get_client_by_node(&client, node_name.clone()).await?;
+
+    let virt_disks = fetch_vdisks(node_client.as_ref()).await?;
+
+    let mut all_partitions: FuturesUnordered<_> = virt_disks
+        .iter()
+        .map(|vdisk| {
+            let id = vdisk.id;
+            let handle = client.api_main().clone();
+            tokio::spawn(async move { (id, handle.get_partitions(id).await) })
+        })
+        .collect();
+
+    let metrics = fetch_metrics(node_client.as_ref()).await?;
+    let typed_metrics: TypedMetrics = metrics.clone().into();
+
+    let space = fetch_space_info(node_client.as_ref()).await?;
+    let node_status = fetch_node_status(node_client.as_ref()).await?;
+    let disks = fetch_disks(node_client.as_ref()).await?;
+
+    let res_disks = proccess_disks(&disks, &space, &metrics);
+    let nodes = fetch_nodes(node_client.as_ref()).await?;
+
+    let nodes = nodes
+        .iter()
+        .map(|node| (&node.name, node))
+        .collect::<HashMap<&NodeName, &dto::Node>>();
+
+    let mut proc_disks = HashMap::new();
+    while let Some(disks) = all_disks.next().await {
+        let Ok(Ok(GetDisksResponse::AJSONArrayWithDisksAndTheirStates(disks))) = disks else {
+            tracing::error!("couldn't get disk inforamtion from node");
+            continue;
+        };
+        for disk in disks {
+            proc_disks.insert(disk.name, disk.is_active);
+        }
+    }
+
+    let mut res_partitions = HashMap::new();
+    while let Some(partitions) = all_partitions.next().await {
+        let Ok((id, Ok(GetPartitionsResponse::NodeInfoAndJSONArrayWithPartitionsInfo(partitions)))) =
+            partitions
+        else {
+            // tracing::error!("couldn't get Partition inforamtion from node"); // Too noisy
+            continue;
+        };
+        if let Some(partitions) = partitions.partitions {
+            res_partitions.insert(id, partitions.len());
+        }
+    }
+
+    let virtual_disks = process_vdisks_for_node(
+        client,
+        &virt_disks,
+        node_name,
+        &proc_disks,
+        &nodes,
+        &res_partitions,
+    )
+    .await;
+
+    let mut rps = RPS::new();
+
+    let status = NodeStatus::from_problems(NodeProblem::default_from_metrics(&typed_metrics));
+
+    rps[Operation::Get] = typed_metrics[RawMetricEntry::PearlGetCountRate].value;
+    rps[Operation::Put] = typed_metrics[RawMetricEntry::PearlPutCountRate].value;
+    rps[Operation::Exist] = typed_metrics[RawMetricEntry::PearlExistCountRate].value;
+    rps[Operation::Delete] = typed_metrics[RawMetricEntry::PearlDeleteCountRate].value;
+
+    let result = Json(DetailedNode {
+        name: node_status.name,
+        hostname: node_status.address,
+        vdisks: virtual_disks,
+        status,
+        metrics: DetailedNodeMetrics {
+            rps,
+            alien_count: typed_metrics[RawMetricEntry::BackendAlienCount].value,
+            corrupted_count: typed_metrics[RawMetricEntry::BackendCorruptedBlobCount].value,
+            space: SpaceInfo {
+                total_disk: space.total_disk_space_bytes,
+                free_disk: space.total_disk_space_bytes - space.used_disk_space_bytes,
+                used_disk: space.used_disk_space_bytes,
+                occupied_disk: space.occupied_disk_space_bytes,
+            },
+            cpu_load: typed_metrics[RawMetricEntry::HardwareBobCpuLoad].value,
+            total_ram: typed_metrics[RawMetricEntry::HardwareTotalRam].value,
+            used_ram: typed_metrics[RawMetricEntry::HardwareUsedRam].value,
+            descr_amount: typed_metrics[RawMetricEntry::HardwareDescrAmount].value,
+        },
+        disks: res_disks,
+    });
+    tracing::trace!("send response: {result:?}");
+
+    Ok(result)
 }
 
 async fn get_client_by_node(