+++ backend/src/services/mod.rs
@@ -1,13 +1,14 @@
 mod prelude {
     pub use super::methods::{
-        fetch_configuration, fetch_metrics, fetch_nodes, fetch_vdisks, get_vdisk_by_id,
+        fetch_configuration, fetch_disks, fetch_metrics, fetch_node_status, fetch_nodes,
+        fetch_space_info, fetch_vdisks, get_vdisk_by_id,
     };
     pub use crate::{
         connector::{
             api::{prelude::*, ApiNoContext},
             ClientError,
         },
-        models::api::*,
+        models::{api::*, bob::*},
         prelude::*,
     };
     pub use axum::{
@@ -26,13 +27,12 @@ pub mod auth;
 pub mod methods;
 
 use api::{
-    get_disks_count, get_node_info, get_nodes_count, get_nodes_list, get_rps, get_space,
-    raw_configuration_by_node, raw_metrics_by_node,
+    get_detailed_node_info, get_disks_count, get_node_info, get_nodes_count, get_nodes_list,
+    get_rps, get_space, get_vdisk_info, get_vdisks_list, raw_configuration_by_node,
+    raw_metrics_by_node,
 };
 use auth::{login, logout, require_auth, AuthState, BobUser, HttpBobClient, InMemorySessionStore};
 use prelude::*;
-
-use self::api::{get_vdisk_info, get_vdisks_list};
 
 type BobAuthState = AuthState<
     BobUser,
@@ -58,6 +58,11 @@ pub fn api_router_v1(auth_state: BobAuthState) -> Result<Router<BobAuthState>, R
         .api_route("/nodes/:node_name", &Method::GET, get_node_info)
         .api_route("/vdisks/list", &Method::GET, get_vdisks_list)
         .api_route("/vdisks/:vdisk_id", &Method::GET, get_vdisk_info)
+        .api_route(
+            "/nodes/:node_name/detailed",
+            &Method::GET,
+            get_detailed_node_info,
+        )
         .api_route(
             "/nodes/:node_name/metrics",
             &Method::GET,