+++ frontend/astro.config.mjs
@@ -6,9 +6,6 @@ import tailwind from '@astrojs/tailwind';
 // https://astro.build/config
 export default defineConfig({
     // output: 'server',
-    redirects: {
-        '/': '/dashboard',
-    },
     integrations: [
         react({
             experimentalReactChildren: true,